+++ access_modifiers/tests/test_performance.py
@@ -7,43 +7,43 @@ import unittest
 class PerformanceTest(unittest.TestCase):
     """Performance tests for the access modifiers."""
 
+    def setUp(self):
+        self.number = 10_000  # How often to repeat the performance test with timeit.
+
     def test_create_private_method(self):
         """Test the time it takes to create private methods as compared to methods without access modifier."""
-        number = 1_000
         time_without_modifier = timeit.timeit("""
 class C:
     def private_method(self):
         pass
-""", setup="from access_modifiers import privatemethod", number=number)
+""", setup="from access_modifiers import privatemethod", number=self.number)
 
         time_with_modifier = timeit.timeit("""
 class C:
     @privatemethod
     def private_method(self):
         pass
-""", setup="from access_modifiers import privatemethod", number=number)
+""", setup="from access_modifiers import privatemethod", number=self.number)
         self.assertLess(time_with_modifier, time_without_modifier * 3)
 
     def test_create_protected_method(self):
         """Test the time it takes to create protected methods as compared to methods without access modifier."""
-        number = 1_000
         time_without_modifier = timeit.timeit("""
 class C:
     def public_method(self):
         pass
-""", setup="from access_modifiers import protectedmethod", number=number)
+""", setup="from access_modifiers import protectedmethod", number=self.number)
 
         time_with_modifier = timeit.timeit("""
 class C:
     @protectedmethod
     def protected_method(self):
         pass
-""", setup="from access_modifiers import protectedmethod", number=number)
-        self.assertLess(time_with_modifier, time_without_modifier * 2)
+""", setup="from access_modifiers import protectedmethod", number=self.number)
+        self.assertLess(time_with_modifier, time_without_modifier * 3)
 
     def test_call_private_method(self):
         """Test the time it takes to call private methods as compared to methods without access modifier."""
-        number = 1_000
         time_without_modifier = timeit.timeit("c.public_method()", setup="""
 from access_modifiers import privatemethod
 class C:
@@ -52,7 +52,7 @@ class C:
     def public_method(self):
         self.private_method()
 c = C()
-""", number=number)
+""", number=self.number)
 
         time_with_modifier = timeit.timeit("c.public_method()", setup="""
 from access_modifiers import privatemethod
@@ -63,12 +63,11 @@ class C:
     def public_method(self):
         self.private_method()
 c = C()
-""", number=number)
-        self.assertLess(time_with_modifier, time_without_modifier * 20)
+""", number=self.number)
+        self.assertLess(time_with_modifier, time_without_modifier * 15)
 
     def test_call_protected_method(self):
         """Test the time it takes to call protected methods as compared to methods without access modifier."""
-        number = 1_000
         time_without_modifier = timeit.timeit("c.public_method()", setup="""
 from access_modifiers import protectedmethod
 class C:
@@ -77,7 +76,7 @@ class C:
     def public_method(self):
         self.protected_method()
 c = C()
-""", number=number)
+""", number=self.number)
 
         time_with_modifier = timeit.timeit("c.public_method()", setup="""
 from access_modifiers import protectedmethod
@@ -88,5 +87,5 @@ class C:
     def public_method(self):
         self.protected_method()
 c = C()
-""", number=number)
+""", number=self.number)
         self.assertLess(time_with_modifier, time_without_modifier * 10)